chore: move rules filtering to rules.go
