+++ secrets/rules/rule.go
@@ -8,6 +8,11 @@ import (
 	"github.com/zricethezav/gitleaks/v8/detect"
 )
 
+type Rule struct {
+	Rule config.Rule
+	Tags []string
+}
+
 // Copied from https://github.com/gitleaks/gitleaks/blob/463d24618fa42fc7629dc30c9744ebe36c5df1ab/cmd/generate/config/rules/rule.go
 func validate(r config.Rule, truePositives []string, falsePositives []string) *config.Rule {
 	// normalize keywords like in the config package
